+++ sentry/src/main/java/io/sentry/util/SampleRateUtil.java
@@ -15,7 +15,7 @@ public final class SampleRateUtil {
       return allowNull;
     }
 
-    return !(sampleRate.isNaN() || (sampleRate > 1.0 || sampleRate <= 0.0));
+    return !(sampleRate.isNaN() || sampleRate > 1.0 || sampleRate <= 0.0);
   }
 
   public static boolean isValidTracesSampleRate(@Nullable Double tracesSampleRate) {
@@ -24,10 +24,17 @@ public final class SampleRateUtil {
 
   public static boolean isValidTracesSampleRate(
       @Nullable Double tracesSampleRate, boolean allowNull) {
-    if (tracesSampleRate == null) {
+    return isValidRate(tracesSampleRate, allowNull);
+  }
+
+  public static boolean isValidProfilesSampleRate(@Nullable Double profilesSampleRate) {
+    return isValidRate(profilesSampleRate, true);
+  }
+
+  private static boolean isValidRate(final @Nullable Double rate, final boolean allowNull) {
+    if (rate == null) {
       return allowNull;
     }
-
-    return !(tracesSampleRate.isNaN() || (tracesSampleRate > 1.0 || tracesSampleRate < 0.0));
+    return !rate.isNaN() && rate >= 0.0 && rate <= 1.0;
   }
 }